updated migrationManager.cs with FundPrices data seeding

+++ Ludens.Models/Entities/Ludens/FundPrices.cs
@@ -6,6 +6,6 @@
         public string FundId { get; set; }
         public Funds Funds { get; set; }
         public DateTime Date { get; set; }
-        public decimal Close { get; set; }
+        public double Close { get; set; }
     }
 }
